refactored sidebar into its own component
