+++ apps/frontend/src/containers/root.css
@@ -96,6 +96,11 @@
 }
 
 /* Navigation Item - Default State */
+a.sidebar-nav-link {
+  text-decoration: none;
+  color: inherit;
+}
+
 .sidebar-nav-item {
   display: flex;
   flex-direction: column;
